requirements.txt
fastapi>=0.95
uvicorn[standard]>=0.22.0
pandas
python-dotenv
SQLAlchemy
PyMySQL
cryptography
google-genai
pydantic-settings
rapidfuzz